平面加了个顶点，阴影正确了，透视矫正插值结果更准了？？？

+++ SoftRendering/Math/Vector2.h
@@ -61,6 +61,6 @@ public:
 
 	inline float length()
 	{
-		return sqrtf(pow(_x, 2) + pow(_y, 2));
+		return sqrtf(powf(_x, 2) + powf(_y, 2));
 	}
 };
